refactor: MakeTableFragmentをActivityに変更
