+++ android-project/MakaseteChoice/app/src/main/java/com/teambrack/makasetechoice/di/ApplicationComponent.kt
@@ -1,10 +1,7 @@
 package com.teambrack.makasetechoice.di
 
 import com.teambrack.makasetechoice.MakaseteChoiceApplication
-import com.teambrack.makasetechoice.di.modules.MainActivityModule
-import com.teambrack.makasetechoice.di.modules.MakeGroupActivityModule
-import com.teambrack.makasetechoice.di.modules.RepositoryModule
-import com.teambrack.makasetechoice.di.modules.ViewModelModule
+import com.teambrack.makasetechoice.di.modules.*
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjector
@@ -17,6 +14,7 @@ import javax.inject.Singleton
         AndroidSupportInjectionModule::class,
         MainActivityModule::class,
         MakeGroupActivityModule::class,
+        MakeTableActivityModule::class,
         RepositoryModule::class,
         ViewModelModule::class
     ]
