+++ android-project/MakaseteChoice/app/src/main/java/com/teambrack/makasetechoice/di/modules/MakeTableActivityModule.kt
@@ -1,8 +1,10 @@
 package com.teambrack.makasetechoice.di.modules
 
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
 import com.teambrack.makasetechoice.di.ViewModelKey
-import com.teambrack.makasetechoice.maketable.MakeTableFragment
+import com.teambrack.makasetechoice.makegroup.MakeGroupActivity
+import com.teambrack.makasetechoice.maketable.MakeTableActivity
 import com.teambrack.makasetechoice.maketable.MakeTableViewModel
 import dagger.Binds
 import dagger.Module
@@ -10,17 +12,19 @@ import dagger.android.ContributesAndroidInjector
 import dagger.multibindings.IntoMap
 
 @Module
-abstract class MakeTableFragmentModule {
+abstract class MakeTableActivityModule {
     @ContributesAndroidInjector(
         modules = [
-            MakeTableFragmentSubModule::class
+            MakeTableActivitySubModule::class
         ]
     )
-    abstract fun contributeMakeTableFragment(): MakeTableFragment
+    abstract fun contributeMakeTableActivity(): MakeTableActivity
 }
 
 @Module
-abstract class MakeTableFragmentSubModule {
+abstract class MakeTableActivitySubModule {
+    @Binds
+    abstract fun providesAppCompatActivity(makeTableActivity: MakeTableActivity): AppCompatActivity
     @Binds
     @IntoMap
     @ViewModelKey(MakeTableViewModel::class)
