+++ android-project/MakaseteChoice/app/src/main/res/layout/activity_make_table.xml
