--- android-project/MakaseteChoice/app/src/main/res/layout/fragment_home.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <variable
-                type="com.teambrack.makasetechoice.home.HomeViewModel"
-                name="viewModel"/>
-    </data>
-    <androidx.constraintlayout.widget.ConstraintLayout
-            android:orientation="vertical"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/make_table"
-                android:id="@+id/make_table_button"
-                android:layout_marginTop="8dp"
-                android:onClick="@{() -> viewModel.onClickMakeTable()}"
-                app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
-                android:layout_marginStart="8dp" app:layout_constraintEnd_toEndOf="parent"
-                android:layout_marginEnd="8dp"
-                app:layout_constraintBottom_toTopOf="@+id/make_group_button"
-                app:layout_constraintVertical_chainStyle="packed"/>
-
-        <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/make_group"
-                android:id="@+id/make_group_button"
-                android:onClick="@{() -> viewModel.onClickMakeGroup()}"
-                app:layout_constraintStart_toStartOf="parent"
-                android:layout_marginStart="8dp" app:layout_constraintEnd_toEndOf="parent"
-                android:layout_marginEnd="8dp"
-                android:layout_marginBottom="8dp" app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/make_table_button" android:layout_marginTop="8dp"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
